add orgstrap.el from tangle
orgstrap.el should not be edited directly, it is automatically tangled
when editing orgstrap.org. The only reason it is being committed is
because it is easier to make orgstrap accessible via melpa if
orgstrap.el is tangled. Also only commit orgstrap.el when it is time
for a release or update cycle.