orgstrap-mode renamed orgstrap-edit-mode and add new orgstrap-mode
Since no changes to the file local variables need to be made, and due
to fact that the new mode needs quite a bit of testing, this is going
it its own development branch. I am fairly impressed that this worked
at all, hack-local-variables hooks are really useful hacks!

Added a change log section and a link to it from the introduction.