note that we need to determine orgstrap-mode behavior when activated
If orgstrap-mode is activated manually it is necessary to determine
whether the orgstrap block has already been run for the current buffer
based on the presence of the eval local variable. We will assume that
the local variable was run and message the user to that effect so that
they can manually run orgstrap-run-block if they know that it has not
been run.