move bootstrap.el code into init.el
As far as I can tell there is no technical reason why the bootstrap
code for use-package needs to be in its own file (e.g. I imagined that
maybe it mattered for the eval-when-compile form). However after
testing against >=emacs-24 without encountering any issues I think it
is ok to move the code into init.el. I'm not 100% sure that the issue
described in #1 applies in the context of init.el and bootstrap.el,
since they are not used with orgstrap.el at all (confusingly), but are
instead intented to be used as a lightweight way to configure a
reproducible environment for exporting org files to html in an
automated pipeline.

Closes #1.