handle setting org-confirm-babel-evaluate in orgstrap block
First pass attempt at handling the use case for being able to set
org-confirm-babel-evaluate in the orgstrap block in a way that allows
those changes to persist.

Since this presents a major avenue for arbitrary code execution, I
think it needs to simmer for a while. I think that it is probably a
better idea to try to roll this into a function that can check the
orgstrap block for a variety of known issues. It is likely to be more
compact and can warn about particularly risky issues beforehand.