further notes on the problem of package manager diversity
I have a feeling that this is an example of a kind of universal
problem that will appear over and over in different contexts. There
are an enormous diversity of ways of getting from point a to point b,
given nothing more than a name as a starting point e.g. orgstrap, and
a simple test e.g. (require 'orgstrap) does not error, as an ending
point. It is trivial to specify the critiera for success, however
implementing the path in a way that works across multiple environments
seems like a step that very often can only be implemented after the
runtime environment is known. For example, imagine that you try to use
an orgstrap file 10 years in the future. Does melpa exist? Who knows!?
You might have to go hunt down a copy of the library that was printed
out before the great disaster and type it in manually, who knows.
This seems quite similar to the issue of containerized reproducibility
vs maximum portability. I think the right solution is going to be to
create a command like orgstrap-projectize or orgstrap-materialize that
will take an orgstrap specification and use it to build a guranteed
reproductible version with everything that it needs bound together.