reval.org save-window-excursion doesn't save-excursion apparently
There is a case where if you have two emacs windows open pointing to
the same buffer in a small enough window then the original indow from
which reval-update-simple is run may also be the window in which
reaval-audit is displayed, if this happens, then deletion and
insertion step happen at the cursor location in the second buffer
window rather than the position of the buffer where the command was
originally run (this is probably a bug in emacs). It seems that
wrapping the window excursion in save-excursion prevents this.