shebang update shebang block to correctly handle passing args
Following a very helpful prompt from Max Nikulin I reviewed the
portability and correctness of how we were passing $args to emacs in
the shebang block setup code.

As a result I have revised the block, and while the changes appear to
be minor, the end result is that args are now correctly passed to
emacs. I have added a full explication of every line of the supporting
polyglot shell code to clarify exactly what is going on at each step.

In summary, never try to assign $@ to a another variable in a posix
shell script because the behavior can be whatever the shell wants.
Passing "${@}" quoted produces consistent behavior across shells.